+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/cds-canvas.component.ts
@@ -16,7 +16,7 @@ import { Intent, Form } from 'src/app/models/intent-model';
 import { Button, Action} from 'src/app/models/action-model';
 
 // UTILS //
-import { INTENT_COLORS, RESERVED_INTENT_NAMES, TYPE_INTENT_ELEMENT, TYPE_OF_MENU, INTENT_TEMP_ID, OPTIONS } from '../../utils';
+import { INTENT_COLORS, RESERVED_INTENT_NAMES, TYPE_INTENT_ELEMENT, TYPE_OF_MENU, INTENT_TEMP_ID, OPTIONS, STAGE_SETTINGS } from '../../utils';
 import { LOGOS_ITEMS } from './../../utils-resources';
 
 
@@ -26,6 +26,7 @@ import { LoggerService } from 'src/chat21-core/providers/abstract/logger.service
 import { TYPE_ACTION } from 'src/app/chatbot-design-studio/utils-actions';
 import { AppStorageService } from 'src/chat21-core/providers/abstract/app-storage.service';
 import { storage } from 'firebase';
+import { LogService } from 'src/app/services/log.service';
 
 // const swal = require('sweetalert');
 
@@ -43,7 +44,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   @Input() onHeaderTestItOut: Observable<Intent>
 
   /** listners */
-  listenerConnectorDrawn: (e: CustomEvent) => void;
+  listnerConnectorDrawn: (e: CustomEvent) => void;
   listnerMovedAndScaled: (e: CustomEvent) => void;
   listnerKeydown: (e: any) => void;
   listnerConnectorSelected: (e: CustomEvent) => void;
@@ -105,6 +106,10 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   testitOutFirstClick: boolean = false;
   IS_OPEN_PANEL_WIDGET: boolean = false;
 
+  /** panel widget loaded */
+  private subscriptionWidgetLoaded: Subscription;
+
+
   /** panel options */
   private subscriptionUndoRedo: Subscription;
   stateUndoRedo: any = {undo:false, redo: false};
@@ -122,6 +127,10 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   IS_OPEN_COLOR_MENU: boolean = false;
   positionColortMenu: any = { 'x': 0, 'y': 0 };
 
+  /** panel widget log */
+  IS_OPEN_WIDGET_LOG: boolean = false;
+  
+
   private logger: LoggerService = LoggerInstance.getInstance()
   
   IS_OPEN_PANEL_INTENT_DETAIL: boolean = false;
@@ -137,7 +146,8 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     public dashboardService: DashboardService,
     private changeDetectorRef: ChangeDetectorRef,
     private route: ActivatedRoute, 
-    public appStorageService: AppStorageService
+    public appStorageService: AppStorageService,
+    public logService: LogService
   ) {
     this.setSubscriptions();
     this.setListnerEvents();
@@ -167,13 +177,16 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     if (this.subscriptionOpenWidgetPanel) {
       this.subscriptionOpenWidgetPanel.unsubscribe();
     }
-    if (this.subscriptionUndoRedo) {
-      this.subscriptionUndoRedo.unsubscribe();
+    if (this.subscriptionWidgetLoaded) {
+      this.subscriptionWidgetLoaded.unsubscribe();
     }
     if (this.subscriptionUndoRedo) {
       this.subscriptionUndoRedo.unsubscribe();
     }
-    document.removeEventListener("connector-drawn", this.listenerConnectorDrawn, false);
+    if (this.subscriptionUndoRedo) {
+      this.subscriptionUndoRedo.unsubscribe();
+    }
+    document.removeEventListener("connector-drawn", this.listnerConnectorDrawn, false);
     document.removeEventListener("moved-and-scaled", this.listnerMovedAndScaled, false);
     document.removeEventListener("start-dragging", this.listnerStartDragging, false);
     document.removeEventListener("keydown", this.listnerKeydown, false);
@@ -191,11 +204,16 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   ngAfterViewInit() {
     this.logger.log("[CDS-CANVAS]  •••• ngAfterViewInit ••••");
     this.stageService.initializeStage(this.id_faq_kb);
-    // this.stageService.initStageSettings(this.id_faq_kb);
+    if(this.stageService.settings?.open_intent_list_state != null){
+      this.IS_OPEN_INTENTS_LIST = this.stageService.settings.open_intent_list_state;
+    }
     
+    
+    // this.stageService.initStageSettings(this.id_faq_kb);
     this.stageService.setDrawer();
     this.connectorService.initializeConnectors();
     this.changeDetectorRef.detectChanges();
+    
     setTimeout(() => {
       this.showStageForLimitTime();
     }, 20000);
@@ -421,6 +439,13 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       }
     });    
 
+    /** SUBSCRIBE TO THE LOADED WIDGET */
+    this.subscriptionWidgetLoaded = this.logService.BSWidgetLoaded.subscribe((event: any) => {
+      this.logger.log("[CDS-CANVAS] logService loaded ", event);
+      if(event){
+        this.IS_OPEN_WIDGET_LOG = true;
+      }
+    });    
   }
 
    /** initialize */
@@ -438,7 +463,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       const numIntents = Object.values(this.mapOfIntents).length;
       const numConnectors = Object.values(this.mapOfConnectors).length;
       this.totElementsOnTheStage = numIntents+numConnectors;
-      this.logger.log('[CDS-CANVAS3] totElementsOnTheStage ::', this.stageService.loaded, numIntents, numConnectors);
+      this.logger.log('[CDS-CANVAS] totElementsOnTheStage ::', this.stageService.loaded, numIntents, numConnectors);
       // scaleAndcenterStageOnCenterPosition(this.listOfIntents)
     }
     this.subscriptionOpenWidgetPanel = this.intentService.BStestiTout.pipe(skip(1)).subscribe((event) => this.onTestItOut(event));
@@ -457,6 +482,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   /** closeAllPanels */
   private closeAllPanels(){
     this.IS_OPEN_PANEL_WIDGET = false;
+    this.IS_OPEN_WIDGET_LOG = false;
     this.IS_OPEN_PANEL_ACTION_DETAIL = false;
     this.IS_OPEN_PANEL_INTENT_DETAIL = false;
     this.IS_OPEN_PANEL_BUTTON_CONFIG = false;
@@ -519,12 +545,12 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   setListnerEvents(){
     /** LISTENER OF TILEDESK STAGE */
     /** triggers when a connector is drawn on the stage  */
-    this.listenerConnectorDrawn = (e: CustomEvent) => {
-    const connector = e.detail.connector;
-    this.setConnectorColor(connector);
-    this.checkAllConnectors(connector);
-  };
-  document.addEventListener("connector-drawn", this.listenerConnectorDrawn, false);
+    this.listnerConnectorDrawn = (e: CustomEvent) => {
+      const connector = e.detail.connector;
+      this.setConnectorColor(connector);
+      this.checkAllConnectors(connector);
+    };
+    document.addEventListener("connector-drawn", this.listnerConnectorDrawn, false);
     
 
     
@@ -714,7 +740,6 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
       }
     };
     document.addEventListener("keydown", this.listnerKeydown, false);
-
   }
 
   // ---------------------------------------------------------
@@ -899,6 +924,7 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
     this.logger.log('[CDS-CANVAS] onToogleSidebarIntentsList  ')
     this.IS_OPEN_INTENTS_LIST = !this.IS_OPEN_INTENTS_LIST;
     this.removeConnectorDraftAndCloseFloatMenu();
+    this.stageService.saveSettings(this.id_faq_kb, STAGE_SETTINGS.openIntentListState, this.IS_OPEN_INTENTS_LIST);
     this.logger.log('[CDS-CANVAS] onToogleSidebarIntentsList   this.IS_OPEN_INTENTS_LIST ',  this.IS_OPEN_INTENTS_LIST)
   }
 
@@ -1242,30 +1268,26 @@ export class CdsCanvasComponent implements OnInit, AfterViewInit{
   // - button configuration panel  
   // -------------------------------------------------------
   onTestItOut(intent: Intent) {
-    // this.testItOut.emit(true);
+    // // this.testItOut.emit(true);
+    this.closeAllPanels();
     this.testitOutFirstClick = true;
     this.logger.log('[CDS-CANVAS] onTestItOut intent ', intent);
-    this.IS_OPEN_PANEL_WIDGET = true
-    this.intentService.startTestWithIntent(intent)
-    // if(typeof event === "boolean"){
-    //   this.IS_OPEN_PANEL_WIDGET = true;
-    // } else {
-    //   this.IS_OPEN_PANEL_WIDGET = !this.IS_OPEN_PANEL_WIDGET;
-    // }
-    if(this.IS_OPEN_PANEL_WIDGET){
-      this.controllerService.closeActionDetailPanel();
-      this.controllerService.closeButtonPanel();
-      // this.intentService.setLiveActiveIntent(null);
-      this.controllerService.closeAddActionMenu();
-      this.connectorService.removeConnectorDraft();
-    }
+    this.intentService.startTestWithIntent(intent);
+    this.controllerService.closeActionDetailPanel();
+    this.controllerService.closeButtonPanel();
+    // // this.intentService.setLiveActiveIntent(null);
+    this.controllerService.closeAddActionMenu();
+    this.connectorService.removeConnectorDraft();
     if(intent){
-      // this.intentSelected = intent;
+      // // this.intentSelected = intent;
       this.intentService.setIntentSelectedById(intent.intent_id);
       this.intentService.setIntentSelected(intent.intent_id);
       this.closeExtraPanels();
     }
 
+    setTimeout(() => {
+      this.IS_OPEN_PANEL_WIDGET = true;
+    }, 500);
   }
 
   /** onActionDeleted */