+++ CHANGELOG.md
@@ -12,6 +12,17 @@
 
 # 1.28.0
 
+# 1.28.0-rc.2
+- **added**: added OpenIntentListState in local storage
+- **added**: added widget log panel
+- **changed**: new panel to change alpha color
+
+# 1.28.0-rc.1
+- **changed**: added sat-popover on change alpha color
+- **bug-fixed**: element position start only the first time
+- **bug-fixed**: cleanup of "stageService" file
+- **bug-fixed**: on live test of chatbot or intent, save scale and position
+
 # 1.27.0-rc.20
 - **bug-fixed**: initStageSettings
 - **bug-fixed**: set default alpha_connectors